Refactor TopMarketsRepository
- rename to MarketTopCoinsRepository
- move to 'topcoins' package
